refactor: rename models to types
