+++ libs/angular-google-charts/src/index.ts
@@ -8,10 +8,10 @@ export * from './lib/components/google-chart/google-chart.component';
 export * from './lib/components/chart-wrapper/chart-wrapper.component';
 export * from './lib/components/dashboard/dashboard.component';
 export * from './lib/components/control-wrapper/control-wrapper.component';
-export * from './lib/helpers/chart.helper';
-export * from './lib/models/events.model';
-export * from './lib/models/chart-type.model';
-export * from './lib/models/control-type.model';
 export * from './lib/components/chart-base/chart-base.component';
-export * from './lib/script-loader/script-loader.service';
+export * from './lib/helpers/chart.helper';
+export * from './lib/types/events';
+export * from './lib/types/chart-type';
+export * from './lib/types/control-type';
+export * from './lib/services/script-loader.service';
 export * from './lib/google-charts.module';
