+++ libs/angular-google-charts/src/lib/components/dashboard/dashboard.component.spec.ts
@@ -2,11 +2,11 @@ import { SimpleChange } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { EMPTY, of, Subject } from 'rxjs';
 
-import { ScriptLoaderService } from '../../script-loader/script-loader.service';
+import { ScriptLoaderService } from '../../services/script-loader.service';
 
 import { DashboardComponent } from './dashboard.component';
 
-jest.mock('../../script-loader/script-loader.service');
+jest.mock('../../services/script-loader.service');
 
 const visualizationMock = {
   Dashboard: jest.fn(),
