+++ libs/angular-google-charts/src/lib/google-charts.module.ts
@@ -5,9 +5,8 @@ import { ChartWrapperComponent } from './components/chart-wrapper/chart-wrapper.
 import { ControlWrapperComponent } from './components/control-wrapper/control-wrapper.component';
 import { DashboardComponent } from './components/dashboard/dashboard.component';
 import { GoogleChartComponent } from './components/google-chart/google-chart.component';
-import { GoogleChartsConfig } from './models/google-charts-config.model';
-import { GOOGLE_CHARTS_CONFIG } from './models/injection-tokens.model';
-import { ScriptLoaderService } from './script-loader/script-loader.service';
+import { ScriptLoaderService } from './services/script-loader.service';
+import { GoogleChartsConfig, GOOGLE_CHARTS_CONFIG } from './types/google-charts-config';
 
 @NgModule({
   declarations: [GoogleChartComponent, ChartWrapperComponent, DashboardComponent, ControlWrapperComponent, ChartEditorComponent],
