+++ libs/angular-google-charts/src/lib/types/chart-type.ts
