+++ libs/angular-google-charts/src/lib/types/control-type.ts
