+++ libs/angular-google-charts/src/lib/types/events.ts
