+++ libs/angular-google-charts/src/lib/types/formatter.ts
@@ -1,0 +1,4 @@
+export interface Formatter {
+  formatter: google.visualization.DefaultFormatter;
+  colIndex: number;
+}
