+++ libs/angular-google-charts/src/lib/types/role.ts
